Update function folder structure
