+++ code/function/fastapi/api/v1/api_v1.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter
-from function.api.v1.endpoints import heartbeat, sample
+from fastapi.api.v1.endpoints import heartbeat, sample
 
 api_v1_router = APIRouter()
 api_v1_router.include_router(sample.router, prefix="/sample", tags=["sample"])
