+++ code/function/fastapi/api/v1/endpoints/sample.py
@@ -1,8 +1,8 @@
 from typing import Any
 
 from fastapi import APIRouter
-from function.models.sample import SampleRequest, SampleResponse
-from function.utils import setup_logging
+from fastapi.models.sample import SampleRequest, SampleResponse
+from fastapi.utils import setup_logging
 
 logger = setup_logging(__name__)
