+++ code/function/fastapi/main.py
@@ -1,0 +1,27 @@
+from fastapi import FastAPI
+from fastapi.api.v1.api_v1 import api_v1_router
+from fastapi.core.config import settings
+
+
+def get_app() -> FastAPI:
+    app = FastAPI(
+        title=settings.PROJECT_NAME,
+        version=settings.APP_VERSION,
+        openapi_url="/openapi.json",
+        debug=settings.DEBUG,
+    )
+    app.include_router(api_v1_router, prefix=settings.API_V1_STR)
+    return app
+
+
+app = get_app()
+
+
+@app.on_event("startup")
+async def startup_event():
+    pass
+
+
+@app.on_event("shutdown")
+async def shutdown_event():
+    pass
