+++ code/function/wrapper/__init__.py
@@ -1,6 +1,6 @@
 import azure.functions as func
-from function.main import fastapi_app
+from fastapi.main import app
 
 
 async def main(req: func.HttpRequest, context: func.Context) -> func.HttpResponse:
-    return await func.AsgiMiddleware(fastapi_app).handle_async(req, context)
+    return await func.AsgiMiddleware(app).handle_async(req, context)
